server/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==3.12.0
cryptography==3.4.7
python-dotenv==0.19.0 